Allow the two gtkD branches to be installed side by side.
When installing GtkD with the make file the headers will be installed in
a gtkd-x subdir in the $(prefix)/include/d, all the libraries are
postfixed with the major version number as follows: libgtkd-2.a.

To help with compiling gtkD applications the Make file now also installs
pkg config files. These can be used as followes:

$dmd myapp.d `pkg-config --cflags --libs gtkd-2`