Remove the detection of the GTK_BASEPATH for Windows.
On my system when i try to use the GTK_BASEPATH instead of adding the
Gtk+ install to my Path, i gat errors from Gtk+ itself because it cant
load some dll's. Presumably because it's looking for them in the Path.
This whould prevent anyone form using this option instead of adding
there Gtk+ install to their Path.