Set the dub target type to none.
With the target type as library, dub will build one library containing
everything, and the separate libraries for each sub package, when the
main target is used as a dependency.

This is not the intended behavior, and can lead to issues as reported
in #202.

dub should probably just build the sub packages when running `dub build`
on the main target.

Reverts: e151e0492d8bcd9f47689878ef3cd48a4f6bc2a3
See Also: #175 and #202