Remove the SimpleGL demo.
The gl demos should be ported to the gtk.GLArea and GtkApplication.

The problem for this demo is that gdk will create an GL context with the
3.2 profile, which disables the deprecated opengl apis used in this
demo.

As a user we can't manualy set the gl profile to because gdk will also
use this internally.

The core demo is a good replacement for this demo.