build: Use Meson instead of custom GNUMake
The custom GNUMake is not really the best experience for developers,
distributors and contributors.

Build systems that are well known are CMake, Autotools, and fortunately
lately Meson was created, which is a secular step in handling builds and
is being adopted by the majority of well known Linux FOSS projects.

This commit adds Meson as the main build system.