Use our own version of Scoped to avoid issues with `this`.
Th Scoped in phobos allocates the payload on the stack. And we pass the
Scoped struct from the callback to the delegate, this moves the Scoped
struct changing the value of `this`.

This causes problems with add/remove toggle_ref in the constructor and
destructor, because the value of the data passed should match in both
calls. And the callbacks used by ObjectG will receive a pointer to the
object before it was moved.

See also: #220