Add a destroy notify function to glib.Idle.
This so we always reset the idle id when the source is destroyed on the
GTK side. For example when the callback returns false, previously we
would always call g_source_remove in the destructor. If the callback
returned false the id passed to g_source_remove is not valid or might
have been reused.